管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -42,7 +42,7 @@ public class EmployeeAction extends ActionBase {
     public void index() throws ServletException, IOException {
 
         //管理者かどうかのチェック //追記
-        //if (checkAdmin()) { //追記
+        if (checkAdmin()) { //追記
 
             //指定されたページ数の一覧画面に表示するデータを取得
             int page = getPage();
@@ -66,7 +66,7 @@ public class EmployeeAction extends ActionBase {
             //一覧画面を表示
             forward(ForwardConst.FW_EMP_INDEX);
 
-        //} //追記
+        } //追記
 
     }
 
@@ -78,13 +78,13 @@ public class EmployeeAction extends ActionBase {
     public void entryNew() throws ServletException, IOException {
 
         //管理者かどうかのチェック //追記
-        //if (checkAdmin()) { //追記
+        if (checkAdmin()) { //追記
             putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
             putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
 
             //新規登録画面を表示
             forward(ForwardConst.FW_EMP_NEW);
-        //} //追記
+        } //追記
     }
 
     /**
@@ -95,7 +95,7 @@ public class EmployeeAction extends ActionBase {
     public void create() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        //if (checkAdmin() && checkToken()) { //追記
+        if (checkAdmin() && checkToken()) { //追記
 
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
@@ -134,7 +134,7 @@ public class EmployeeAction extends ActionBase {
                 redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
             }
 
-        //}
+        }
     }
 
     /**
@@ -145,7 +145,7 @@ public class EmployeeAction extends ActionBase {
     public void show() throws ServletException, IOException {
 
         //管理者かどうかのチェック //追記
-        //if (checkAdmin()) { //追記
+        if (checkAdmin()) { //追記
 
             //idを条件に従業員データを取得する
             EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -161,7 +161,7 @@ public class EmployeeAction extends ActionBase {
 
             //詳細画面を表示
             forward(ForwardConst.FW_EMP_SHOW);
-        //}
+        }
 
     }
 
@@ -173,7 +173,7 @@ public class EmployeeAction extends ActionBase {
     public void edit() throws ServletException, IOException {
 
         //管理者かどうかのチェック //追記
-        //if (checkAdmin()) { //追記
+        if (checkAdmin()) { //追記
 
             //idを条件に従業員データを取得する
             EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -191,7 +191,7 @@ public class EmployeeAction extends ActionBase {
             //編集画面を表示する
             forward(ForwardConst.FW_EMP_EDIT);
 
-        //} //追記
+        } //追記
     }
 
     /**
@@ -202,7 +202,7 @@ public class EmployeeAction extends ActionBase {
     public void update() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        //if (checkAdmin() && checkToken()) { //追記
+        if (checkAdmin() && checkToken()) { //追記
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -238,7 +238,7 @@ public class EmployeeAction extends ActionBase {
                 //一覧画面にリダイレクト
                 redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
             }
-        //}
+        }
     }
 
     /**
@@ -249,7 +249,7 @@ public class EmployeeAction extends ActionBase {
     public void destroy() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        //if (checkAdmin() && checkToken()) { //追記
+        if (checkAdmin() && checkToken()) { //追記
 
             //idを条件に従業員データを論理削除する
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -259,7 +259,7 @@ public class EmployeeAction extends ActionBase {
 
             //一覧画面にリダイレクト
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
-        //}
+        }
     }
 
     /**
